Create V5 - Sad Balls Active
My final version. The Sad Balls now move around independantly, rebounding off the walls (all 4 of them) and each other. The red ball rebounds off of them as well however does not affect the Sad Ball trajectory. In the future (if I'm arsed), I will make the Sad Balls rebound off of the Smiley Ball and the paddle.

Probably got lots of bugs in it but only a draft. I've left in most of original, incorrect coding in comments for myself to look over in case I want to try a new method I can see if I already tried it before.